Telepresence robot code
The telepresence robot is designed to create a link between an actual classroom setting and the remote students and lecturers who are unable to attend classes physically, by audial and visual feedback and interaction. It is operated through a smartphone and it comprises connectivity through the ESP8266 microcontroller, servos for the head and motors for motion.